+++ libblkid/src/superblocks/ntfs.c
@@ -149,10 +149,10 @@ static int probe_ntfs(blkid_probe pr, const struct blkid_idmag *mag)
 	off = le64_to_cpu(ns->mft_cluster_location) * sector_size *
 		sectors_per_cluster;
 
-	DBG(LOWPROBE, ul_debug("NTFS: sector_size=%d, mft_record_size=%d, "
-			"sectors_per_cluster=%d, nr_clusters=%ju "
-			"cluster_offset=%jd",
-			(int) sector_size, mft_record_size,
+	DBG(LOWPROBE, ul_debug("NTFS: sector_size=%"PRIu16", mft_record_size=%"PRIu32", "
+			"sectors_per_cluster=%"PRIu32", nr_clusters=%"PRIu64" "
+			"cluster_offset=%"PRIu64"",
+			sector_size, mft_record_size,
 			sectors_per_cluster, nr_clusters,
 			off));
